Move createReturns to separate module
